+++ model/utils.py
@@ -1,4 +1,6 @@
 import mlflow
+import dvc.api
+import pandas as pd
 
 
 def yield_artifacts(run_id, path=None):
@@ -24,3 +26,15 @@ def fetch_logged_data(run_id):
         "tags": tags,
         "artifacts": artifacts,
     }
+    
+def dvc_open(path, url, branch):
+    with dvc.api.open(
+        path = path,  ## 데이터 경로
+        repo = url,  ## github repo 경로,
+        rev =  branch ## 현재는 branch 기준
+    ) as f:
+        return pd.read_csv(f, sep=",")
+
+def find_experiment_id(experiment_name):
+    current_experiment = dict(mlflow.get_experiment_by_name(experiment_name))
+    return current_experiment["experiment_id"]
